download comments

+++ main.py
@@ -6,7 +6,7 @@ import requests
 from urllib.parse import urljoin, urlparse
 
 from pathvalidate import sanitize_filename
-from requests.models import HTTPError, Response
+from requests.exceptions import HTTPError
 from bs4 import BeautifulSoup
 
 
@@ -18,14 +18,17 @@ def check_for_redirect(response):
         raise HTTPError
 
 
-def parse_book_page(id):
-    url = f'https://tululu.org/b{id}'
+def parse_book_page(book_id):
+    url = f'https://tululu.org/b{book_id}/'
     response = requests.get(url)
     response.raise_for_status()
+    check_for_redirect(response)
     soup = BeautifulSoup(response.text, 'lxml')
     title, author = (el.strip() for el in soup.find('h1').text.split('::'))
-    img = urljoin(url, soup.find('div', class_='bookimage').find('img')['src'])
-    return title, img
+    cover_url = urljoin(url, soup.find('div', class_='bookimage').find('img')['src'])
+    comments = [comment.text.split(')')[1] for comment in soup.find_all('div', class_='texts')]
+    logger.info(f'Parse book page {book_id}')
+    return title, cover_url, comments
 
 
 def make_filename(id, title, folder='books'):
@@ -33,47 +36,48 @@ def make_filename(id, title, folder='books'):
     return os.path.join(folder, filename)
 
 
-def load_cover(url, folder='books'):
+def download_cover(cover_url, folder='books'):
     filename = os.path.join(
         folder,
-        urlparse(url).path.split('/')[-1]
+        urlparse(cover_url).path.split('/')[-1]
     )
+    if not os.path.isfile(filename):
+        response = requests.get(cover_url)
+        response.raise_for_status()
+        logger.info(f'Download cover {filename}')
+        with open(filename, 'wb') as file:
+            file.write(response.content)
+
+
+def download_book(folder, filename, id):
+    url = f'https://tululu.org/txt.php?id={id}'
     response = requests.get(url)
     response.raise_for_status()
     with open(filename, 'wb') as file:
         file.write(response.content)
-
-
-def load_book(id):
-    url = f'https://tululu.org/txt.php?id={id}'
-    response = requests.get(url)
-    response.raise_for_status()
-    try:
-        check_for_redirect(response)
-        title, img = parse_book_page(id)
-        load_cover(img)
-        filename = make_filename(id, title)
-        with open(filename, 'wb') as file:
-            file.write(response.content)
-        logger.info(f'Save book {filename} {img}')
-    except HTTPError:
-        pass
+    logger.info(f'Download book {filename}')
 
 
 def main():
     logging.basicConfig(
-        level=logging.ERROR,
+        level=logging.INFO,
         format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-    )
-    logger.setLevel(logging.INFO)    
+    )    
     folder = os.path.join(
         os.path.dirname(__file__),
         'books'
         )
     os.makedirs(folder, exist_ok=True)
-    for id in range(1, 11):
-        load_book(id)
+    for book_id in range(1, 11):
+        try:
+            title, cover_url, comments = parse_book_page(book_id)
+            print(title, '\n'.join(comments))
+            # filename = make_filename(book_id, title, folder)
+            # download_cover(cover_url, folder)
+            # download_book(folder, filename, book_id)
 
+        except HTTPError:
+            logger.info(f'No book for {book_id}')
 
 if __name__ == '__main__':
     main()
